backend/requirements.txt
fastapi[standard]==0.115.5
pandas==2.2.3
torch==2.2.2
torchvision==0.17.2
numpy==1.26.1
scikit-learn==1.5.2
ultralytics==8.3.49
pillow==11.0.0
mlflow==2.18.0
prometheus-fastapi-instrumentator==7.0.0
evidently==0.7.16
transformers==4.40.0
accelerate==0.30.0